requirements.txt
streamlit
uuid
python-docx
sqlite3